requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
youtube-transcript-api==0.6.1
requests==2.31.0
